refactor(cwl): keep simple recipient API
