+++ internal/routes/register.go
@@ -145,9 +145,6 @@ func Register(app *fiber.App, a apptypes.Deps, wrap func(fiber.Handler) fiber.Ha
 	app.Post("/v2/roster/views/preview", rosterRead(previewRosterView(a)))
 	app.Post("/v2/roster/membership-changes/validate", rosterRead(validateRosterMembershipProposal(a)))
 	app.Post("/v2/roster/membership-changes", rosterWrite(applyRosterMembershipChanges(a)))
-	app.Get("/v2/cwl/bonus-awards/context", authCWLBonusQueryRead(a, wrap, getCWLBonusContext(a)))
-	app.Post("/v2/cwl/bonus-awards", authCWLBonusCreate(a, wrap, submitCWLBonusAwards(a)))
-	app.Get("/v2/cwl/bonus-awards/history", authCWLBonusQueryRead(a, wrap, getCWLBonusHistory(a)))
 	app.Get("/v2/server/:server_id/cwl/:clan_tag/bonus-recipients", serverRead(getCWLBonusRecipients(a)))
 	app.Put("/v2/server/:server_id/cwl/:clan_tag/bonus-recipients", serverWrite(replaceCWLBonusRecipients(a)))
 	app.Get("/v2/roster/metrics", rosterRead(listRosterMetrics(a)))
