+++ internal/routes/register_test.go
@@ -150,9 +150,6 @@ func TestRegisterOmitsRemovedRoutesAndKeepsV2Routes(t *testing.T) {
 		"/v2/clan/:clan_tag/legend-history",
 		"/v2/legends/history/:season",
 		"/v2/player/:player_tag/cwl/history",
-		"/v2/cwl/bonus-awards/context",
-		"/v2/cwl/bonus-awards",
-		"/v2/cwl/bonus-awards/history",
 		"/v2/server/:server_id/cwl/:clan_tag/bonus-recipients",
 		"/v2/cwl/:clan_tag/seasons",
 		"/v2/cwl/:clan_tag",
